Revert "Changed lastLongitude type back to DECIMAL"
This reverts commit 0275726fccc5fc0afe1ab117e61e7c31da489d14.

+++ src/models/User.js
@@ -27,10 +27,10 @@ class User extends Model {
             company: DataTypes.STRING,
             position: DataTypes.STRING,
             about: DataTypes.STRING,
-            lastLongitude: DataTypes.DECIMAL(15, 15),
-            lastLatitude: DataTypes.DECIMAL(15, 15),
-            // lastLongitude: DataTypes.REAL,
-            // lastLatitude: DataTypes.REAL,
+            // lastLongitude: DataTypes.DECIMAL(10, 6),
+            // lastLatitude: DataTypes.DECIMAL(10, 6),
+            lastLongitude: DataTypes.REAL,
+            lastLatitude: DataTypes.REAL,
             lastTimeSuperLikeWasUsed: DataTypes.DATE
         }, {
             sequelize,//conexão
